adds comments to MSPortfolioViewController
