+++ MockStock/Controllers/MSPortfolioViewController.swift
@@ -27,7 +27,7 @@ class MSPortfolioViewController: UIViewController {
         return v
     }()
     
-    // MARK: Methods
+    // MARK: ViewController Methods
     override func viewDidLoad() {
         super.viewDidLoad()
         
@@ -43,14 +43,14 @@ class MSPortfolioViewController: UIViewController {
         // Add header views to the root view
         view.addSubview(networkActivityIndicator)
         
-        // add collection view
+        // Add and configure collection view
         view.addSubview(collectionView)
         collectionView.delegate = self
         collectionView.dataSource = self
         collectionView.register(MSPortfolioItemCell.self, forCellWithReuseIdentifier: "PortfolioItem")
-        collectionView.register(PortfolioMetaInfoCell.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: "header")
+        collectionView.register(MSPortfolioMetaInfoCell.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: "header")
         
-        // collection view constraints
+        // Collection view constraints
         collectionView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 0).isActive = true
         collectionView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: 0).isActive = true
         collectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 0).isActive = true
@@ -71,6 +71,15 @@ class MSPortfolioViewController: UIViewController {
         fetchData()
     }
     
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        MSPortfolioData.sharedInstance.items.removeAll()
+        MSPortfolioData.sharedInstance.buyingPower = 0
+        MSPortfolioData.sharedInstance.portfolioValue = 0
+        collectionView.reloadData()
+    }
+    
+    // MARK: API Methods
     func fetchData() {
         // 0. Start activity indicator animation
         networkActivityIndicator.startAnimating()
@@ -82,11 +91,14 @@ class MSPortfolioViewController: UIViewController {
         }
         
         // 2. Send portfolio data request to server using authentication token
+        // 2.1 - Configure HTTP request
         let urlString = "\(MSAPI.baseUrl)/api/portfolio"
         guard let url = URL(string: urlString) else { return }
         var urlRequest = URLRequest(url: url)
         urlRequest.httpMethod = "GET"
         urlRequest.addValue(token, forHTTPHeaderField: "Authorization")
+        
+        // 2.2 - Send & response to HTTP request
         URLSession.shared.dataTask(with: urlRequest) { [weak self] (data, response, error) in
             if let e = error { print(e) }
             guard let data = data else { return }
@@ -122,25 +134,22 @@ class MSPortfolioViewController: UIViewController {
             } catch let jsonErr {
                 print(jsonErr)
             }
+            
+            // Clean up UI after response recieved
             DispatchQueue.main.async {
                 self?.collectionView.reloadData()
                 self?.networkActivityIndicator.stopAnimating()
             }
         }.resume() // fires the session
     }
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        MSPortfolioData.sharedInstance.items.removeAll()
-        MSPortfolioData.sharedInstance.buyingPower = 0
-        MSPortfolioData.sharedInstance.portfolioValue = 0
-        collectionView.reloadData()
-    }
+    
     @objc func logOut() {
         present(MSLoginViewController(), animated: false, completion: nil)
     }
 
 }
 
+// MARK: UICollectionView Data Source implementation
 extension MSPortfolioViewController: UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
     func numberOfSections(in collectionView: UICollectionView) -> Int {
         return 1
@@ -178,7 +187,7 @@ extension MSPortfolioViewController: UICollectionViewDataSource, UICollectionVie
     }
     
     func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        let headerCell = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "header", for: indexPath) as! PortfolioMetaInfoCell
+        let headerCell = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "header", for: indexPath) as! MSPortfolioMetaInfoCell
         headerCell.networthValue.text = String(format: "$%.02f", MSPortfolioData.sharedInstance.buyingPower + MSPortfolioData.sharedInstance.portfolioValue)
         headerCell.buyingpowerValue.text = String(format: "$%.02f", MSPortfolioData.sharedInstance.buyingPower)
         headerCell.portfolioValue.text = String(format: "$%.02f", MSPortfolioData.sharedInstance.portfolioValue)
@@ -196,100 +205,4 @@ extension MSPortfolioViewController: UICollectionViewDataSource, UICollectionVie
         }
     }
     
-}
-
-class PortfolioMetaInfoCell: UICollectionViewCell {
-    var networth: UILabel = {
-        let l = UILabel()
-        l.text = "Net Worth:"
-        l.font = UIFont(name: "HelveticaNeue-Thin", size: 22)
-        l.textColor = .gray
-        l.translatesAutoresizingMaskIntoConstraints = false
-        return l
-    }()
-    var networthValue: UILabel = {
-        let l = UILabel()
-        l.text = "$13,000.42"
-        l.font = UIFont(name: "HelveticaNeue-Thin", size: 22)
-        l.textColor = .gray
-        l.translatesAutoresizingMaskIntoConstraints = false
-        l.textAlignment = .right
-        return l
-    }()
-    var buyingPower: UILabel = {
-        let l = UILabel()
-        l.text = "Buying Power:"
-        l.font = UIFont(name: "HelveticaNeue-Thin", size: 22)
-        l.textColor = .gray
-        l.translatesAutoresizingMaskIntoConstraints = false
-        return l
-    }()
-    var buyingpowerValue: UILabel = {
-        let l = UILabel()
-        l.text = "$9,001.34"
-        l.font = UIFont(name: "HelveticaNeue-Thin", size: 22)
-        l.textColor = .gray
-        l.translatesAutoresizingMaskIntoConstraints = false
-        l.textAlignment = .right
-        return l
-    }()
-    var portfolio: UILabel = {
-        let l = UILabel()
-        l.text = "Portfolio Value:"
-        l.font = UIFont(name: "HelveticaNeue-Thin", size: 22)
-        l.textColor = .gray
-        l.translatesAutoresizingMaskIntoConstraints = false
-        return l
-    }()
-    var portfolioValue: UILabel = {
-        let l = UILabel()
-        l.text = "$3,999.00"
-        l.font = UIFont(name: "HelveticaNeue-Thin", size: 22)
-        l.textColor = .gray
-        l.translatesAutoresizingMaskIntoConstraints = false
-        l.textAlignment = .right
-        return l
-    }()
-    
-    override func didMoveToSuperview() {
-        addSubview(networth)
-        addSubview(networthValue)
-        addSubview(buyingPower)
-        addSubview(buyingpowerValue)
-        addSubview(portfolio)
-        addSubview(portfolioValue)
-        
-        let leftInset = CGFloat(30.0)
-        let rightInset = CGFloat(-30.0)
-        
-        networth.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor, constant: 10).isActive = true
-        networth.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: leftInset).isActive = true
-        networth.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.centerXAnchor, constant: 0).isActive = true
-        networth.heightAnchor.constraint(equalToConstant: 25).isActive = true
-        
-        networthValue.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor, constant: 10).isActive = true
-        networthValue.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.centerXAnchor, constant: 0).isActive = true
-        networthValue.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: rightInset).isActive = true
-        networthValue.heightAnchor.constraint(equalToConstant: 25).isActive = true
-        
-        buyingPower.topAnchor.constraint(equalTo: networth.bottomAnchor, constant: 10).isActive = true
-        buyingPower.leadingAnchor.constraint(equalTo: leadingAnchor, constant: leftInset).isActive = true
-        buyingPower.trailingAnchor.constraint(equalTo: centerXAnchor, constant: 0).isActive = true
-        buyingPower.heightAnchor.constraint(equalToConstant: 25).isActive = true
-        
-        buyingpowerValue.topAnchor.constraint(equalTo: networth.bottomAnchor, constant: 10).isActive = true
-        buyingpowerValue.leadingAnchor.constraint(equalTo: centerXAnchor, constant: 0).isActive = true
-        buyingpowerValue.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: rightInset).isActive = true
-        buyingpowerValue.heightAnchor.constraint(equalToConstant: 25).isActive = true
-        
-        portfolio.topAnchor.constraint(equalTo: buyingPower.bottomAnchor, constant: 10).isActive = true
-        portfolio.leadingAnchor.constraint(equalTo: leadingAnchor, constant: leftInset).isActive = true
-        portfolio.trailingAnchor.constraint(equalTo: centerXAnchor, constant: 0).isActive = true
-        portfolio.heightAnchor.constraint(equalToConstant: 25).isActive = true
-        
-        portfolioValue.topAnchor.constraint(equalTo: buyingPower.bottomAnchor, constant: 10).isActive = true
-        portfolioValue.leadingAnchor.constraint(equalTo: centerXAnchor, constant: 0).isActive = true
-        portfolioValue.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: rightInset).isActive = true
-        portfolioValue.heightAnchor.constraint(equalToConstant: 25).isActive = true
-    }
 }
